循環参照修正
